requirements.txt
asgiref==3.5.0
astroid==2.9.3
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.10
colorclass==2.2.2
cryptography==36.0.1
defusedxml==0.7.1
Django==4.0.2
django-admin-thumbnails==0.2.6
django-ckeditor==6.2.0
django-cors-headers==3.13.0
django-environ==0.8.1
django-js-asset==1.2.2
django-mptt==0.13.4
djangorestframework==3.14.0
docopt==0.6.2
environ==1.0
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
oauthlib==3.2.0
packaging==21.3
pep8==1.7.1
pip-upgrader==1.4.15
pkg_resources==0.0.0
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.7
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
terminaltables==3.1.10
toml==0.10.2
typing_extensions==4.4.0
urllib3==1.26.8
wrapt==1.13.3
